More unsigned long -> uintptr_t fixes.

+++ cpu-exec.c
@@ -781,7 +781,7 @@ void cpu_x86_frstor(CPUX86State *s, target_ulong ptr, int data32)
    the effective address of the memory exception. 'is_write' is 1 if a
    write caused the exception and otherwise 0'. 'old_set' is the
    signal set which should be restored */
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -831,7 +831,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined(TARGET_ARM)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -869,7 +869,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
     return 1;
 }
 #elif defined(TARGET_SPARC)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -907,7 +907,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
     return 1;
 }
 #elif defined (TARGET_PPC)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -957,7 +957,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined(TARGET_M68K)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -996,7 +996,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined (TARGET_MIPS)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -1046,7 +1046,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined (TARGET_MICROBLAZE)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -1096,7 +1096,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined (TARGET_SH4)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -1141,7 +1141,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 }
 
 #elif defined (TARGET_ALPHA)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -1185,7 +1185,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
     return 1;
 }
 #elif defined (TARGET_CRIS)
-static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
+static inline int handle_cpu_signal(uintptr_t pc, uintptr_t address,
                                     int is_write, sigset_t *old_set,
                                     void *puc)
 {
@@ -1234,7 +1234,7 @@ static inline int handle_cpu_signal(unsigned long pc, unsigned long address,
 #if defined(__APPLE__)
 # include <sys/ucontext.h>
 
-# define EIP_sig(context)  (*((unsigned long*)&(context)->uc_mcontext->ss.eip))
+# define EIP_sig(context)  (*((uintptr_t*)&(context)->uc_mcontext->ss.eip))
 # define TRAP_sig(context)    ((context)->uc_mcontext->es.trapno)
 # define ERROR_sig(context)   ((context)->uc_mcontext->es.err)
 # define MASK_sig(context)    ((context)->uc_sigmask)
@@ -1259,7 +1259,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #else
     struct ucontext *uc = puc;
 #endif
-    unsigned long pc;
+    uintptr_t pc;
     int trapno;
 
 #ifndef REG_EIP
@@ -1270,7 +1270,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #endif
     pc = EIP_sig(uc);
     trapno = TRAP_sig(uc);
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              trapno == 0xe ?
                              (ERROR_sig(uc) >> 1) & 1 : 0,
                              &MASK_sig(uc), puc);
@@ -1299,7 +1299,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
                        void *puc)
 {
     siginfo_t *info = pinfo;
-    unsigned long pc;
+    uintptr_t pc;
 #ifdef __NetBSD__
     ucontext_t *uc = puc;
 #elif defined(__OpenBSD__)
@@ -1309,7 +1309,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #endif
 
     pc = PC_sig(uc);
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              TRAP_sig(uc) == 0xe ?
                              (ERROR_sig(uc) >> 1) & 1 : 0,
                              &MASK_sig(uc), puc);
@@ -1384,7 +1384,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
     if (TRAP_sig(uc) != 0x400 && (DSISR_sig(uc) & 0x02000000))
         is_write = 1;
 #endif
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write, &uc->uc_sigmask, puc);
 }
 
@@ -1415,7 +1415,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 	is_write = 1;
     }
 
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write, &uc->uc_sigmask, puc);
 }
 #elif defined(__sparc__)
@@ -1430,15 +1430,15 @@ int cpu_signal_handler(int host_signum, void *pinfo,
     uint32_t *regs = (uint32_t *)(info + 1);
     void *sigmask = (regs + 20);
     /* XXX: is there a standard glibc define ? */
-    unsigned long pc = regs[1];
+    uintptr_t pc = regs[1];
 #else
 #ifdef __linux__
     struct sigcontext *sc = puc;
-    unsigned long pc = sc->sigc_regs.tpc;
+    uintptr_t pc = sc->sigc_regs.tpc;
     void *sigmask = (void *)sc->sigc_mask;
 #elif defined(__OpenBSD__)
     struct sigcontext *uc = puc;
-    unsigned long pc = uc->sc_pc;
+    uintptr_t pc = uc->sc_pc;
     void *sigmask = (void *)(long)uc->sc_mask;
 #endif
 #endif
@@ -1471,7 +1471,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 	break;
       }
     }
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write, sigmask, NULL);
 }
 
@@ -1482,7 +1482,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
 #if (__GLIBC__ < 2 || (__GLIBC__ == 2 && __GLIBC_MINOR__ <= 3))
@@ -1492,7 +1492,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 #endif
     /* XXX: compute is_write */
     is_write = 0;
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write,
                              &uc->uc_sigmask, puc);
 }
@@ -1504,13 +1504,13 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
     pc = uc->uc_mcontext.gregs[16];
     /* XXX: compute is_write */
     is_write = 0;
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write,
                              &uc->uc_sigmask, puc);
 }
@@ -1526,7 +1526,7 @@ int cpu_signal_handler(int host_signum, void *pinfo, void *puc)
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long ip;
+    uintptr_t ip;
     int is_write = 0;
 
     ip = uc->uc_mcontext.sc_ip;
@@ -1544,7 +1544,7 @@ int cpu_signal_handler(int host_signum, void *pinfo, void *puc)
       default:
 	  break;
     }
-    return handle_cpu_signal(ip, (unsigned long)info->si_addr,
+    return handle_cpu_signal(ip, (uintptr_t)info->si_addr,
                              is_write,
                              &uc->uc_sigmask, puc);
 }
@@ -1556,13 +1556,13 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     siginfo_t *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
     pc = uc->uc_mcontext.psw.addr;
     /* XXX: compute is_write */
     is_write = 0;
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write, &uc->uc_sigmask, puc);
 }
 
@@ -1578,7 +1578,7 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 
     /* XXX: compute is_write */
     is_write = 0;
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr,
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr,
                              is_write, &uc->uc_sigmask, puc);
 }
 
@@ -1589,13 +1589,13 @@ int cpu_signal_handler(int host_signum, void *pinfo,
 {
     struct siginfo *info = pinfo;
     struct ucontext *uc = puc;
-    unsigned long pc;
+    uintptr_t pc;
     int is_write;
 
     pc = uc->uc_mcontext.sc_iaoq[0];
     /* FIXME: compute is_write */
     is_write = 0;
-    return handle_cpu_signal(pc, (unsigned long)info->si_addr, 
+    return handle_cpu_signal(pc, (uintptr_t)info->si_addr, 
                              is_write,
                              &uc->uc_sigmask, puc);
 }
